requirements.txt
pandas
numpy<1.20.0
pyldavis==2.1.2
nltk
wordcloud
gensim
xlrd
openpyxl
scikit-learn
python-dotenv
flair
spark-nlp==2.7.4
sentence-transformers
umap-learn
hdbscan
textblob
jupyter
ipykernel
notebook
